+++ ext_emconf.php
@@ -12,10 +12,10 @@ $EM_CONF[$_EXTKEY] = [
     'constraints' => [
         'depends' => [
             'php' => '7.0.0-7.9.99',
-            'typo3' => '8.7.0-8.7.99',
-            'frontend' => '8.7.0-8.7.99',
-            'fluid' => '8.7.0-8.7.99',
-            'fluid_styled_content' => '8.7.0-8.7.99',
+            'typo3' => '8.7.0-9.5.99',
+            'frontend' => '8.7.0-9.5.99',
+            'fluid' => '8.7.0-9.5.99',
+            'fluid_styled_content' => '8.7.0-9.5.99',
         ],
         'conflicts' => [
         ],